Add lsp semantic token highlighting (disabled)
Semantic token highlighting uses the lsp to highlight tokens by their
semantic type.

Enabling this feature might slow down text scrolling. The slowdown is
caused by the font locking mechanism in Emacs.

For now, I'm adding this setting as a comment. In the future I will
investigate the slowness, and if a proper solution will be
found (e.g. removing the bold weights from the relevant faces), this
will be turned on.