xref: Fix history not tracked until the first jump back
xref is an internal Emacs package, however, it is also maintained as a
public one. Some packages are using the public version directly,
creating a discrepancy in package initialization.  By forcing the
public package to load before the private one is used we avoid
multiple initializations and we solve the issue at hand.